Python/RealPython/walrus/requirements.txt
blessings==1.7
bpython==0.21
certifi==2021.5.30
charset-normalizer==2.0.4
curtsies==0.3.5
cwcwidth==0.1.4
feedparser==6.0.8
greenlet==1.1.1
html2text==2020.1.16
idna==3.2
importlib-resources==5.2.2
parse==1.19.0
Pygments==2.9.0
pyxdg==0.27
realpython-reader==1.0.0
requests==2.26.0
sgmllib3k==1.0.0
six==1.16.0
typing==3.7.4.3
urllib3==1.26.6
zipp==3.5.0
